redis/booking_service/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
email-validator==2.2.0

alembic==1.13.3
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9

grpcio==1.67.1
grpcio-tools==1.67.1
protobuf==5.28.3